refactor(types,network,storage,agent): 协议枚举定型家族——CompressionType 下沉 + 值域校验 + PeerRpcWireStatus/PeerStreamDirection/MonitorSampleKind 定型
审计批次 C 第二程 P1-1 / P1-2 / P2-5 / P2-6（同批文件物理交织，原子提交）：

P1-1 CompressionType 全链定型：
- 枚举自 storage/cpp/compressor.h 下沉 common/types/cpp/compression_type.h，
  底层类型 int8_t → uint8_t 定型（0-3 补码表示相同，盘面/线上 1 字节编码
  不变）；storage 侧经 include 复用全局命名空间同名定义，既有消费者零改动
- ObjectHeader.compression_type_ 定型（落盘对象头，字节级红线）：
  serialize/serialize_trailer 的 memcpy 路径 sizeof 恒 1 字节；
  deserialize/deserialize_trailer 改 uint8_t 临时直读 + is_valid_compression_type
  值域校验先于类型化赋值（越界 = 损坏，确定性拒绝）
- ChunkSource::compression_type() 返回 int → CompressionType；memory/disk/
  network 三实现同步（-1 未知哨兵消除：失败场景 failed() 权威，NONE 为死值；
  NetworkChunkSource 原 0→-1 映射删除——-1 与 NONE 在 make_block_read_pipeline
  同落 DecompressStage(nullptr) 直通，行为等价）
- DataResponseMessage.chunk_compression_type_ / PeerStreamStartMessage.
  compression_type_ 定型（network 不再以裸数值规避 storage 依赖）
- MessageProtocol::decode 补值域校验（对齐 is_valid_message_type 先例）：
  DataResponseMessage/PeerStreamStartMessage 解码后拒绝越界压缩类型
- decompressing_streambuf / database / data_service / data_server / fly_stream
  全链裸 static_cast 消除；FlyStream::sink_comp_/sink_effective_compression 定型

P1-2 PeerRpcWireStatus 签名定型：
- send_response/ResponseHandler 的 uint8_t status → PeerRpcWireStatus；
  wire status 字节在 server_loop 直读处单点 cast，出口 send_response 组帧处
  单点 cast；BYE 比较/notify_failure/send_not_ready/send_bye/handle_bye 及
  worker_agent 响应路由 switch 的反向 cast 全部消除（越界 wire 值仍由
  消费端 switch default 归类真失败，行为不变）

P2-5 PeerStreamDirection 新枚举（REQUEST=0/RESPONSE=1）：
- PeerStreamStartMessage.direction_ / PeerStreamRxState.direction 定型；
  send_stream_start/send_stream_payload/PeerStreamWriter 构造参数定型；
  server_loop 魔法比较 m.direction_ == 0 消除

P2-6 MonitorSampleKind 新枚举（PERIODIC=0/EVENT=1，network 侧——monitor
依赖 network，反向定义违层）：
- MonitorSample.kind_ 定型；worker_agent/master_agent 事件采样与
  MetricsDb 落库消费点同步

字节级测试锚定：
- object_header_test：CompressionFieldIsSingleByteWireValue（fixed 段末字节
  逐值 0x00-0x03 断言 + sizeof==1）、TrailerCompressionByteMatchesFixedByte、
  内存/trailer 双路径越界拒绝负例、is_valid_compression_type 边界
- message_protocol_test：PeerStreamStart round-trip 字节布局（direction 1B +
  comp 1B 直通）、decode 越界拒绝两例（stream/META）

全量 C++ 单测 113/113 通过。

+++ src/monitor/cpp/metrics_db.cpp
@@ -333,7 +333,7 @@ void MetricsDb::record_worker_samples(uint64_t worker_id, const CMVector<Monitor
                           sqlite3_bind_int64(s, 8, sp.host_load1_x100_);
                           sqlite3_bind_int64(s, 9, static_cast<sqlite3_int64>(sp.net_read_bytes_));
                           sqlite3_bind_int64(s, 10, static_cast<sqlite3_int64>(sp.net_write_bytes_));
-                          sqlite3_bind_int(s, 11, sp.kind_);
+                          sqlite3_bind_int(s, 11, static_cast<int>(sp.kind_));
                       });
         }
     });
